requirements.txt
fastapi==0.124.4
langchain==1.1.3
langchain_core==1.2.0
langchain_google_genai==4.0.0
pydantic==2.12.5
python-dotenv==1.2.1
uvicorn[standard]
pyyaml>=6.0
